+++ flink-python/pyflink/examples/datastream/connectors/kafka_csv_format.py
@@ -20,7 +20,7 @@ import sys
 
 from pyflink.common import Types, JsonRowDeserializationSchema, CsvRowSerializationSchema
 from pyflink.datastream import StreamExecutionEnvironment
-from pyflink.datastream.connectors import FlinkKafkaProducer, FlinkKafkaConsumer
+from pyflink.datastream.connectors.kafka import FlinkKafkaProducer, FlinkKafkaConsumer
 
 
 # Make sure that the Kafka cluster is started and the topic 'test_csv_topic' is
