+++ flink-python/pyflink/examples/datastream/connectors/pulsar.py
@@ -21,9 +21,9 @@ import sys
 
 from pyflink.common import SimpleStringSchema, WatermarkStrategy
 from pyflink.datastream import StreamExecutionEnvironment
-from pyflink.datastream.connectors import PulsarSource, PulsarSink, PulsarSerializationSchema, \
-    StartCursor, StopCursor, SubscriptionType, PulsarDeserializationSchema, DeliveryGuarantee, \
-    TopicRoutingMode
+from pyflink.datastream.connectors.pulsar import PulsarSource, PulsarSink, DeliveryGuarantee, \
+    StartCursor, StopCursor, SubscriptionType, PulsarDeserializationSchema, TopicRoutingMode, \
+    PulsarSerializationSchema
 
 if __name__ == '__main__':
     logging.basicConfig(stream=sys.stdout, level=logging.INFO, format="%(message)s")
